+++ src/js/video.js
@@ -3,6 +3,7 @@
  */
 import document from 'global/document';
 import * as setup from './setup';
+import * as stylesheet from './utils/stylesheet.js';
 import Component from './component';
 import EventTarget from './event-target';
 import Player from './player';
@@ -93,6 +94,16 @@ var videojs = function(id, options, ready){
   // If not, set up a new player and return the instance.
   return tag['player'] || new Player(tag, options, ready);
 };
+
+// Add default styles
+let style = stylesheet.createStyleElement('vjs-styles-defaults');
+let head = document.querySelector('head');
+head.insertBefore(style, head.firstChild);
+stylesheet.setTextContent(style, `
+  .video-js {
+    width: 300px;
+    height: 150px;
+`);
 
 // Run Auto-load players
 // You have to wait at least once in case this script is loaded after your video in the DOM (weird behavior only with minified version)