+++ src/js/tech/html5.js
@@ -93,7 +93,7 @@ class Html5 extends Tech {
     let emulatedTt = this.textTracks();
 
     // remove native event listeners
-    if (tt) {
+    if (tt && tt.removeEventListener) {
       tt.removeEventListener('change', this.handleTextTrackChange_);
       tt.removeEventListener('addtrack', this.handleTextTrackAdd_);
       tt.removeEventListener('removetrack', this.handleTextTrackRemove_);
@@ -184,7 +184,7 @@ class Html5 extends Tech {
   proxyNativeTextTracks_() {
     let tt = this.el().textTracks;
 
-    if (tt) {
+    if (tt && tt.addEventListener) {
       tt.addEventListener('change', this.handleTextTrackChange_);
       tt.addEventListener('addtrack', this.handleTextTrackAdd_);
       tt.addEventListener('removetrack', this.handleTextTrackRemove_);