+++ CHANGELOG.md
@@ -95,6 +95,7 @@ CHANGELOG
 * @mmcc made sure controls respect muted attribute ([view](https://github.com/videojs/video.js/pull/2408))
 * @dmlap switched global options back to an object at videojs.options ([view](https://github.com/videojs/video.js/pull/2461))
 * @ogun fixed a typo in the Turkish translation ([view](https://github.com/videojs/video.js/pull/2460))
+* @gkatsev fixed text track errors on dispose and in cross-browser testing ([view](https://github.com/videojs/video.js/pull/2466))
 
 --------------------