+++ CHANGELOG.md
@@ -99,6 +99,7 @@ CHANGELOG
 * @mmcc added type=button to button components ([view](https://github.com/videojs/video.js/pull/2471))
 * @mmcc Fixed IE by using setAttribute to set &#x27;type&#x27; property ([view](https://github.com/videojs/video.js/pull/2487))
 * @misternoneill fixed vertical slider issues ([view](https://github.com/videojs/video.js/pull/2469))
+* @gkatsev moved default and player dimensions to style els at the top of HEAD ([view](https://github.com/videojs/video.js/pull/2482))
 
 --------------------