+++ CHANGELOG.md
@@ -102,6 +102,7 @@ CHANGELOG
 * @gkatsev moved default and player dimensions to style els at the top of HEAD ([view](https://github.com/videojs/video.js/pull/2482))
 * @gkatsev moved default and player dimensions to style els at the top of HEAD el ([view](https://github.com/videojs/video.js/pull/2482))
 * @gkatsev removed non-default track auto-disabling ([view](https://github.com/videojs/video.js/pull/2475))
+* @gkatsev exported event helpers on videojs object ([view](https://github.com/videojs/video.js/pull/2491))
 
 --------------------