+++ CHANGELOG.md
@@ -104,6 +104,7 @@ CHANGELOG
 * @gkatsev removed non-default track auto-disabling ([view](https://github.com/videojs/video.js/pull/2475))
 * @gkatsev exported event helpers on videojs object ([view](https://github.com/videojs/video.js/pull/2491))
 * @nickygerritsen fixed texttrack handling in IE10 ([view](https://github.com/videojs/video.js/pull/2481))
+* @gkatsev deep clone el for iOS to preserve tracks ([view](https://github.com/videojs/video.js/pull/2494))
 
 --------------------